+++ changelog.md
@@ -3,6 +3,7 @@
 ## 2.0.0 - 2017-05-17
 
 * Bump up @superbalist/js-pubsub to ^2.0.0
+* Add new publishBatch method to GoogleCloudPubSubAdapter
 
 ## 1.0.2 - 2017-05-15